[feat] 뷰모델 분리
