+++ app/src/main/java/org/sopt/certi/presentation/ui/mypage/state/MyPageUiSate.kt
@@ -1,8 +1,16 @@
 package org.sopt.certi.presentation.ui.mypage.state
 
 import org.sopt.certi.core.state.UiState
+import org.sopt.certi.domain.model.user.CertificationCount
 import org.sopt.certi.domain.model.user.MyPageInfo
 
 data class MyPageUiSate(
-    val myPageInfoLoadState: UiState<MyPageInfo> = UiState.Loading
+    val myPageInfoLoadState: UiState<Unit> = UiState.Loading,
+    val myPageInfo: MyPageInfo = MyPageInfo(
+        nickname = "",
+        email = "",
+        profileImageUrl = "",
+        jobs = emptyList(),
+        certificationCount = CertificationCount(0, 0, 0)
+    )
 )
